requirements.txt
bcrypt==3.2.0
fastapi==0.110.0
mysql-connector-python==8.3.0
passlib==1.7.4
pydantic==2.6.4
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.17
sqlalchemy==2.0.29
uvicorn==0.27.1
requests==2.31.0
pytest==7.4.2
starlette==0.36.3
httpx==0.28.0
anyio==4.6.2.post1